BulletPool Added, BulletBase changes
Added BulletPool class that holds a list of bullets which can be fired at a given firePoint and disables rather than destroys them once their duration has reached 0. Its a circular list so we create all bullets at the beginning of the scene and don't instantiate or destroy any afterwards, just set them to active or inactive. I also changed the BulletBase's CleanUp() method to disable, rather than destroy, the bullet. This works well if we use bulletPools for everything. If we for some reason want to destroy a bullet directly we may have to call Destroy(gameObject) explicitly.